Fix: like 생성, 삭제 예외 처리
LikeSerializer
- product_id write_only 필드로 추가

urls.py
- LikeListView와 LikeCreateView 합침

LikeListView
- CreateView와 합쳐서 주석처리함

LikeListCreateView
- ListView를 추가함
- 기존 쿼리 파라미터로 받던 product 정보를 request body로 받음
- Product의 likes가 3씩 증가하는 버그 발생
- Like모델이 먼저 수정되도록 순서 변경

ListDestroyView
- delete를 여러번 했을때 인스턴스가 없는데도 product의 likes가 계속 감소하는 버그 발생
- 이미 삭제된 인스턴스일 경우 에러 발생시키고 동작 수행 못하게 함

+++ apps/like/views.py
@@ -1,7 +1,7 @@
 from django.db import transaction, IntegrityError
 from django.db.models import QuerySet, F
 from rest_framework import generics, permissions
-from rest_framework.exceptions import ValidationError
+from rest_framework.exceptions import ValidationError, NotFound
 
 from apps.like.models import Like
 from apps.like.permissions import IsUserOrReadOnly
@@ -9,28 +9,31 @@ from apps.like.serializers import LikeSerializer
 from apps.product.models import Product
 
 
-class LikeListView(generics.ListAPIView):
+# class LikeListView(generics.ListAPIView):
+#     serializer_class = LikeSerializer
+#     permission_classes = [permissions.IsAuthenticated]
+#
+#     def get_queryset(self) -> QuerySet[Like]:
+#         user = self.request.user
+#         return Like.objects.filter(user=user).order_by("-created_at")
+
+
+class LikeListCreateView(generics.ListCreateAPIView):
     serializer_class = LikeSerializer
-    permission_classes = [permissions.IsAuthenticated]
+    permissions_classes = [permissions.IsAuthenticated, IsUserOrReadOnly]
 
     def get_queryset(self) -> QuerySet[Like]:
         user = self.request.user
         return Like.objects.filter(user=user).order_by("-created_at")
 
-
-class LikeCreateView(generics.CreateAPIView):
-    queryset = Like.objects.all()
-    serializer_class = LikeSerializer
-    permissions_classes = [permissions.IsAuthenticated, IsUserOrReadOnly]
-
     def perform_create(self, serializer: LikeSerializer) -> None:
-        product_id = self.kwargs.get("pk")
-        product = Product.objects.get(pk=product_id)
+        product_id = self.request.data.get("product_id")
+        # product = Product.objects.get(pk=product_id)
 
         try:
             with transaction.atomic():
+                serializer.save(user=self.request.user, product_id=product_id)
                 Product.objects.filter(pk=product_id).update(likes=F("likes") + 1)
-                serializer.save(user=self.request.user, product=product)
                 # product.likes = F("likes") + 1
                 # product.save(update_fields=["likes"])
         except IntegrityError:
@@ -43,12 +46,16 @@ class LikeDestroyView(generics.DestroyAPIView):
 
     def get_object(self) -> QuerySet[Like]:
         product_id = self.kwargs.get("pk")
-        return Like.objects.filter(user=self.request.user, product_id=product_id)
+        like = Like.objects.filter(user=self.request.user, product_id=product_id)
+        if not like:
+            raise NotFound("No Like matches the given query.")
+        return like
 
     @transaction.atomic
     def perform_destroy(self, instance):
         product_id = self.kwargs.get("pk")
-        Product.objects.filter(pk=product_id).update(likes=F("likes") + 1)
+        if instance:
+            instance.delete()
+            Product.objects.filter(pk=product_id).update(likes=F("likes") - 1)
         # product.likes = F("likes") - 1
         # product.save(update_fields=["likes"])
-        instance.delete()
